Add the function of basic operations
For the operator buttons and equal buttons work, classes and event
Listeners were added to them. Users can do basic operations by the GUI.

The function add() was modified so that user can get the correct answer
instead of a string.

The multiply and divide sign in the condition statements in operate()
was changed to be same as the textContent of buttons.